refactor: remove Redux integration and update movie types
- Removed @reduxjs/toolkit and react-redux dependencies from package.json and package-lock.json.
- Deleted moviesSlice.ts and store.ts files to eliminate Redux state management for movies.
- Updated import paths for MoviePropTypes from "@/db" to "@/types" in relevant components.
- Removed unused getAllMovies utility function.
- Cleaned up _app.tsx to remove Redux Provider wrapping.
