+++ components/OtherMovies.tsx
@@ -1,4 +1,4 @@
-import { MoviePropTypes } from "@/db";
+import { MoviePropTypes } from "@/types";
 import Image from "next/image";
 import Link from "next/link";
 import { useRouter } from "next/router";
