+++ components/common/MovieBox.tsx
@@ -1,4 +1,4 @@
-import { MoviePropTypes } from "@/db";
+import { MoviePropTypes } from "@/types";
 import Image from "next/image";
 import Link from "next/link";
 import React, { useState } from "react";
