+++ package.json
@@ -11,7 +11,6 @@
   "dependencies": {
     "@ducanh2912/next-pwa": "^9.7.2",
     "@hookform/resolvers": "^3.3.2",
-    "@reduxjs/toolkit": "^1.9.7",
     "axios": "^1.5.1",
     "firebase": "^12.15.0",
     "i18next": "^23.5.1",
@@ -24,7 +23,6 @@
     "react-hook-form": "^7.48.2",
     "react-i18next": "^13.2.2",
     "react-icons": "^4.11.0",
-    "react-redux": "^8.1.3",
     "react-toastify": "^9.1.3",
     "superjson": "^1.13.3",
     "swiper": "^10.3.1",
