+++ pages/_app.tsx
@@ -3,8 +3,6 @@ import "@/i18n";
 import type { AppProps } from "next/app";
 import { useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { Provider } from "react-redux";
-import { store } from "@/store";
 
 export default function App({ Component, pageProps }: AppProps) {
   const { i18n } = useTranslation();
@@ -14,9 +12,5 @@ export default function App({ Component, pageProps }: AppProps) {
     i18n.changeLanguage(savedLanguage);
   }, []);
 
-  return (
-    <Provider store={store}>
-      <Component {...pageProps} />
-    </Provider>
-  );
+  return <Component {...pageProps} />;
 }
